Implement RegionalHealthCheck and incorporate review feedback.
